Split osgWamIkfast into multiple files
