+++ examples/osgExamples/osgWamIkfast/InputHandler.hpp
@@ -1,0 +1,81 @@
+/*
+ * Copyright (c) 2017, Graphics Lab, Georgia Tech Research Corporation
+ * Copyright (c) 2017, Personal Robotics Lab, Carnegie Mellon University
+ * All rights reserved.
+ *
+ * This file is provided under the following "BSD-style" License:
+ *   Redistribution and use in source and binary forms, with or
+ *   without modification, are permitted provided that the following
+ *   conditions are met:
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *   * Redistributions in binary form must reproduce the above
+ *     copyright notice, this list of conditions and the following
+ *     disclaimer in the documentation and/or other materials provided
+ *     with the distribution.
+ *   THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND
+ *   CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
+ *   INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
+ *   MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ *   DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
+ *   CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ *   SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ *   LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF
+ *   USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
+ *   AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ *   LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ *   ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ *   POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#include <dart/dart.hpp>
+#include <dart/gui/osg/osg.hpp>
+#include <dart/utils/utils.hpp>
+#include <dart/utils/urdf/urdf.hpp>
+
+using namespace dart::dynamics;
+using namespace dart::simulation;
+
+class RelaxedPosture;
+class WamWorld;
+
+class InputHandler : public ::osgGA::GUIEventHandler
+{
+public:
+
+  InputHandler(dart::gui::osg::Viewer* viewer, WamWorld* teleop,
+               const SkeletonPtr& wam, const WorldPtr& world);
+
+  void initialize();
+
+  bool handle(const ::osgGA::GUIEventAdapter& ea,
+              ::osgGA::GUIActionAdapter&) override;
+
+protected:
+
+  dart::gui::osg::Viewer* mViewer;
+
+  WamWorld* mTeleop;
+
+  SkeletonPtr mWam;
+
+  WorldPtr mWorld;
+
+  Eigen::VectorXd mRestConfig;
+
+  std::vector<bool> mConstraintActive;
+
+  std::vector<std::size_t> mEndEffectorIndex;
+
+  std::vector< std::pair<Eigen::Vector6d, Eigen::Vector6d> > mDefaultBounds;
+
+  Eigen::aligned_vector<Eigen::Isometry3d> mDefaultTargetTf;
+
+  std::shared_ptr<RelaxedPosture> mPosture;
+
+  std::shared_ptr<dart::constraint::BalanceConstraint> mBalance;
+
+  char mOptimizationKey;
+
+  std::vector<bool> mMoveComponents;
+};
