+++ examples/osgExamples/osgWamIkfast/WamWorld.hpp
@@ -1,0 +1,84 @@
+/*
+ * Copyright (c) 2017, Graphics Lab, Georgia Tech Research Corporation
+ * Copyright (c) 2017, Personal Robotics Lab, Carnegie Mellon University
+ * All rights reserved.
+ *
+ * This file is provided under the following "BSD-style" License:
+ *   Redistribution and use in source and binary forms, with or
+ *   without modification, are permitted provided that the following
+ *   conditions are met:
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *   * Redistributions in binary form must reproduce the above
+ *     copyright notice, this list of conditions and the following
+ *     disclaimer in the documentation and/or other materials provided
+ *     with the distribution.
+ *   THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND
+ *   CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
+ *   INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
+ *   MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ *   DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
+ *   CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ *   SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ *   LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF
+ *   USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
+ *   AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ *   LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ *   ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ *   POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#include <dart/dart.hpp>
+#include <dart/gui/osg/osg.hpp>
+#include <dart/utils/utils.hpp>
+#include <dart/utils/urdf/urdf.hpp>
+
+using namespace dart::dynamics;
+using namespace dart::simulation;
+
+class WamWorld : public dart::gui::osg::WorldNode
+{
+public:
+
+  enum MoveEnum_t
+  {
+    MOVE_Q = 0,
+    MOVE_W,
+    MOVE_E,
+    MOVE_A,
+    MOVE_S,
+    MOVE_D,
+    MOVE_F,
+    MOVE_Z,
+
+    NUM_MOVE
+  };
+
+  WamWorld(WorldPtr world, SkeletonPtr robot);
+
+  void setMovement(const std::vector<bool>& moveComponents);
+
+  void customPreRefresh() override;
+
+  bool mAmplifyMovement;
+
+protected:
+
+  SkeletonPtr mWam;
+  std::size_t iter;
+
+  EndEffectorPtr l_foot;
+  EndEffectorPtr r_foot;
+
+  EndEffectorPtr l_hand;
+  EndEffectorPtr r_hand;
+
+  std::vector<IK::Analytical::Solution> mSolutions;
+
+  Eigen::VectorXd grad;
+
+  // Order: q, w, e, a, s, d
+  std::vector<bool> mMoveComponents;
+
+  bool mAnyMovement;
+};
